requirements.txt
google-generativeai==0.8.6
pandas==2.2.2
pandas-datareader==0.10.0
pandas-gbq==0.30.0
pandas-stubs==2.2.2.240909
pyngrok==8.0.0
streamlit==1.56.0
